made separate component for posts
